Script no longer restricted to program files
Previously, I made an installer so that the script and the 3 custom tray icons would be automatically placed in the Program Files (x86) of the user, in order for the path to files to remain constant and load the tray icons. But I have since discovered the A_ScriptDir command, so the script will work as long as the icons are in the same folder as it.